Mutable metrics descriptors (#15917)
Replacing the immutable `MetricTagger` with its internal `String`
representation to `MetricDescriptor` and `MutableMetricDescriptor`.
`MutableMetricDescriptor` is used to construct an `MetricDescriptor`
that is passed from the dynamic collection layer to the
`MetricsCollector` instance and to the `MetricsPublisher`s beyond it.

There are two motivations for doing this change:
- The String representation is hard to work with and mutate when needed
(such as when merging member and client side metrics)
- The String representation produces huge amount of garbage both during
collection and when publishing to MC that turned out to be unacceptable
in large scale.
- In many places it made it easier to deal with the metric descriptor
than with the Strings. This change made it possible to remove the String
parser for the JMX exposure and made the code easier to digest around it.

In line with this change, the dynamic collection API changed to the
following:
```
public interface DynamicMetricsProvider {
    void provideDynamicMetrics(Supplier<? extends MutableMetricDescriptor> descriptorSupplier, MetricsCollectionContext context);
}

public interface MetricsCollectionContext {
    void collect(MutableMetricDescriptor descriptor, Object source);
    void collect(MutableMetricDescriptor descriptor, String name, ProbeLevel level, ProbeUnit unit, long value);
    void collect(MutableMetricDescriptor descriptor, String name, ProbeLevel level, ProbeUnit unit, double value);
}
```

This made collection code look like this:
```
MutableMetricDescriptor rootDescriptor = descriptorSupplier
        .get()
        .withPrefix("raft.group");
for (Entry<CPGroupId, RaftNodeMetrics> entry : nodeMetrics.entrySet()) {
    CPGroupId groupId = entry.getKey();
    MutableMetricDescriptor descriptor = rootDescriptor
           .copy()
           .withDiscriminator("groupId", String.valueOf(groupId.getId()))
           .withTag("name", groupId.getName());
    context.collect(descriptor, entry.getValue());
}
```

To further decrease allocation rate a `PoolingMetricDescriptorSupplier`
is introduced exclusively for collecting the dynamic metrics. Once a
metric is collected, its `MetricDescriptor` gets recycled and made
available for reuse. After each collection cycle, the references between
the pooling supplier and the created descriptors are cleared to prevent
leaking if references to a descriptor is stored mistakenly. Since there
is this pool in use, the descriptors supplied to the dynamic metric
providers must not be stored anywhere.

In addition to moving to the mutable descriptor representation the
`MetricCompressor` has changed too. It is now using a dictionary based
delta compression further improving the compression by serializing
through `DeflaterOutputStream` (which was used before).