EntryLoader: ignore entries with calculated TTL equal to 0
When a MetadataAwareValue was loaded (for instance in
DefaultRecordStore.loadRecordOrNull()) exactly (in millis)
at its expiration time, the entry would be marked as
non-expirable. This was because we convert the expiration
time to TTL (expiration time - current time) and if the TTL
was 0, we treated the record as non-expirable.

Changed the above TTL comparison code from greater/less-than
to greater/less-than-or-equal.

Also added missing (internal) javadoc to
RecordStore.putFromLoad[Backup].

Fixes https://github.com/hazelcast/hazelcast/issues/15991