Make gathering multimap stats not to break last access merge policy
Gathering multimap statistics calls
MultiMapPartitionContainer#getMultiMapContainer() that updates the
lastAccessTime of the returned container, which in the end breaks the
last access merge policy. This is fixed by making gathering statistics
not to count an access. Also, retrieving the container on backups
for the given partition made not to be considered an access.

Fixes #16001