Fix failing AnswerTest
Follow-up on #16003 that fixed #16001 by introducing a second parameter
to `MultiMapPartitionContainer#getMultiMapContainer(String,boolean)` to
indicate whether or not the container should be `access()`'ed. This
broke `AnswerTest` and that can't be fixed easily since there is no such
method in HZ 3.11. This is fixed by making the aforementioned method
private and adding back the previously existing
`getMultiMapContainer(String name)` and introducing
`getMultiMapContainerWithoutAccess(String name)` that delegate to the
private `getMultiMapContainer(String,boolean)`.

Fixes #16011