Don't call two processor methods w/o outbox.reset (#1743)
When we called `tryProcessWatermark`, we proceeded to `tryProcess` in a
recursion without calling `outbox.reset()`. This could lead to a situation when
each `tryProcessWatermark` and `process` offer to outbox and are rejected.
Without a `reset()`, the second offer will fail on an assertion. This compelled
us to check for `outbox.hasUnfinishedItem()` when returning from
`tryProcessWatermark()` in `AsyncTransformUsingServiceOrderedP`, and this call
seemed superfluous without knowledge of the inner workings of
`ProcessorTasklet`.

The contract of `outbox.offer()` says that if it returns `false, we should
return from the processor method. But it didn't say that we should return
`false` from the processor method, some methods don't even allow it, such as
`process()`.

The fix was to not recurse in `ProcessorTasklet` after calling
`tryProcessorWatermark()`.