Move JsonSerializable to private API (#16032)
Move JsonSerializable to private API

JsonSerializable is supposed to be used only when exchanging statistics
with management center. Previously, it was public API because some
statistics classes such as LocalMapStats were exposed as public methods.

Now we move JsonSerializable to private API. Because of this, we have to
cast whenever we want to transform from/to JSON as most internal code
exchanges stats interfaces such as LocalMapStats. The alternative is
that the internal API exchanges concrete classes such as
LocalMapStatsImpl but this is not possible since sometimes there are two
different implementations for the same stats interface.

Addresses https://github.com/hazelcast/hazelcast/pull/15888#discussion_r340582364