Refine member metrics configuration (#15818)
This change refines the member-side metrics configuration to address two
issues with the previously added config:
- It was not clear that `retention-seconds` was for MC only
- Further improving the MC and JMX configs was not supported by the
attributes

To address these, the configuration changes as follows.

Before this change:
```
<metrics enabled="true" mc-enabled="true" jmx-enabled="true">
    <collection-interval-seconds>5</collection-interval-seconds>
    <retention-seconds>5</retention-seconds>
    <metrics-for-data-structures>false</metrics-for-data-structures>
    <minimum-level>INFO</minimum-level>
</metrics>
```

After this change:
```
<metrics enabled="true">
    <management-center enabled="true">
        <retention-seconds>5</retention-seconds>
    </management-center>
    <jmx enabled="true"/>
    <collection-frequency-seconds>5</collection-frequency-seconds>
</metrics>
```

The programmatic and YAML configuration follows the same structure. With the defaults above. 

The `metrics-for-data-structures` and `level` settings are removed from 
the configuration and are configurable with the following boolean system 
properties only:
- `hazelcast.metrics.datastructures.enabled`: to disable collecting 
datastructure metrics, enabled by default
- `hazelcast.metrics.debug.enabled`: to collect `DEBUG` level metrics

`MANDATORY` level can no longer be configurable. These settings were 
added to the 3.x line to better control the memory footprint of the collection 
system. Since in the 4.x line the vast majority of the metrics are discovered 
dynamically, the footprint is almost constant. What might change is the size 
of the compressed blob sent to MC (if enabled) and the JMX beans (if enabled). 
The system properties are kept as options to further tune the footprint without 
disabling the metrics system entirely. 

It is also a requirement that all metrics configuration should be
overrideable via system properties. This is also implemented in this
change by overriding the configuration from the properties during the
`Node` instance's creation. The concrete system properties that override
the configuration fields are added to the javadocs of the metrics 
config classes and to the xsd too.

Besides refining the changes above, the following gaps were also closed:
- Improve javadoc and xsd documentation for the metrics config elements
- Add Spring XML configuration
- Add config comments to `hazelcast-full-example.[xml|yaml]`
- Add metrics config to `hazelcast-fullconfig.[xml|yaml]`
- Add metrics metrics config to `hazelcast-default.[xml|yaml]`