Fix XmlConfigImportVariableReplacementTest failing on Windows
Recently the import and variable replacement logic in XML was improved
that introduced using "file:///" URLs for referencing the files to
import. It takes a different path in the ConfigLoader that understands
it as a URL rather a file. The difference between the two paths from the
filing tests point of view is that the file path checks if the file
exists, while the URL path checks only if the URL is correct. In this
specific case the file path throws an InvalidConfigurationException,
because some part of the config (that is to be imported) missing. The
URL path, however, fails with a regular HazelcastException. In the
latter case it cannot be determined if the configuration is invalid or
the resource is unavailable temporarily.

The tests are fixed by removing "file:///" URLs from the affected parts
of the tests, and adding a case that tests specifically the URL path.

Fixes #16054