Attempted fix for WanRESTTest and MapWanSyncRESTTest
The issue we've recently encountered was that wiremock stubbing was
being done on a wrong method. This may happen if, during stubbing, some
other method is called on the mock instance. For example, in the code:
`when(wanServiceMock.syncAllMaps("atob", "B")).thenReturn(expectedUuid)`
if we were to invoke `wanServiceMock.toString()` after the call to the
`when` method but before the call to the `thenReturn` method, the
`expectedUUID` would have been set on the `toString()` method, which is
wrong. This is an inherent issue with wiremock and can be easily
reproduced with debugging in an IDE.

With this fact, we assume there was some call to the mock instance
during stubbing, which caused both tests to fail. We eliminate this
cause by setting up the stubbing before we even start the instance.

Fixes:
https://github.com/hazelcast/hazelcast/issues/16053
https://github.com/hazelcast/hazelcast-enterprise/issues/3364