Fixed flaky HDMapPutAllWithBouncingMemberTest (#16059)
Rewrote the test to use BounceMemberRule with enabled option to gracefully
shutdown a bouncing member. Without graceful shutdown of bouncing member
we may loose some updates of map.putAll().

The old version of the test was suffering from this issue so that if
we terminate (hard shutdown) a member just after an update on the partition owner
(residing on the member being terminated) succeeded, there is a chance that
the partition's replica(s) will never get the update. This is a result of
current "sync" replication design which doesn't fail an operation if an
update hasn't reached all replicas.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/3342