Fix HDEvictionTest (#16045)
When calling map.put, the invocation may return before the backup was
created. In such case, the backup record store will not be created and
the assertion may fail with a NPE.

Added the assertion that the record store is created eventually. Also,
did some JDK8 cleanup and made the test to use small instances, at least
in the case of on-heap.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/3362