Run forced eviction if map has eviction configured.
Issue:
We should run forced eviction only if a map has eviction configured. Otherwise,
it shouldn't run and should throw Native OOME. In current form, forced eviction 
is evicting all entries regardless of the eviction configuration.

Reasoning for the fix:
Effectively we have 1 type of forced eviction strategy and it is based on to 
evict a percentage of entries. So far forced eviction first tries to evict 20 
percent of entries then if it still gets Native OOME, it retries by evicting 
100 percent of entries by using evictAll method.

Changes:
Based on the reasoning, refactored forced eviction code that we only have 
SingleRecordStoreForcedEviction and MultipleRecordStoreForcedEviction and made 
them work with any percentage. By default we have 2 values for percentages: 20 
and 100. Removed other unneeded eviction strategies that are using evictAll.