Fix for AbstractListenersOnReconnectTest
Since CLIENT_CONNECTED event is fired in an executor, the initial
one can be delivered late so that it could be count as `reconnect`,
even though the client is not reconnected yet.

Added the listener in the config to count both initial and second
CONNECTED event to make the test more consistent.

Added `hasNext` check for iterators to make sure eventual assertions
will work correctly.

The log that is added to diagnose the test is converted to Ilogger
to be able to track which test it belongs to.

fixes https://github.com/hazelcast/hazelcast/issues/8254