Remove providing cache near cache metrics on member (#16127)
There are no cache near cache metrics available on the member side.
This change removes the code attempts to provide them.

Some explanation:
CacheStatisticsImpl.getNearCacheStatistics() is clear on that it should
never be called. Added the near cache metrics when copied the related
code from the previously existing StatisticsAwareMetricSet that had the
following

private NearCacheStats getNearCacheStats(LocalInstanceStats localInstanceStats) {
    if (localInstanceStats instanceof LocalMapStatsImpl) {
        LocalMapStats localMapStats = (LocalMapStats) localInstanceStats;
        return localMapStats.getNearCacheStats();
    } else if (localInstanceStats instanceof CacheStatistics) {
        CacheStatistics localMapStats = (CacheStatistics) localInstanceStats;
        return localMapStats.getNearCacheStatistics();
    } else {
        return null;
    }
}

Note that the second branch (the else if) was always a dead branch for
two reasons:
- ICacheService didn't implement StatisticsAwareService interface until
recently, so StatisticsAwareMetricSet never seen cache metrics.
- Since it implemented the interface, its getStats() returns
LocalCacheStat instances, not CacheStatistics instances. Therefore, the
branch was not taken again.

Fixes hazelcast/hazelcast-enterprise#3388