Make StatisticsAwareServices self-registering as metrics providers
Recent change removed StatisticsAwareMetricsSet that was responsible for
providing metrics from these services. This changed to unconditionally
register the implementer services in NodeEngineImpl, which failed if the
cache was not initialized (if the jcache.jar is not present). Therefore,
all services made self-registering. They register themselves in init()
if the datastructure metrics are enabled. There are two exceptions: the
DistributedExecutorService and (Enterprise)WanReplicationService that
doesn't check the value of the datastructure flag.

Also, an NPE is fixed in EnterpriseWanReplicationService's
provider method by checking if the distributed data structures are
present.

Additionally, an assertion is added to fail during the metrics
collection if the collection cycle encounters an exception AND the test
is testing metrics collection specifically by directly calling
MetricRegistry.collect() (otherwise the AssertionError is thrown on a
different thread). This means it is a best-effort assertion. In the
future we might want to introduce a junit rule that reads out this info
from the metrics collection.

Fixes #16136 and hazelcast/hazelcast-enterprise#3391