Fix snapshot chunking (#1771)
We didn't subtract the header length from `usableChunkCapacity`. As a result,
when a single entry was just shy of the `usableChunkCapacity` large and
inserted into an empty chunk, it would be inserted into the current chunk and
not into a separate chunk. Later when flushing, "buffer full" was thrown when
appending the terminator.

Other changes:
- check that the chunkSize is a power of two, the code depends on this
- "put operation overwrote a previous value" assertion was thrown, but we don't
  check the future for an error, we check the `firstError`, it was lost
- replace `whenCompleteAsync` with `whenComplete`, no need to offload, the
  logic is lightweight

Fixes #1769