Checks for existing static config when adding config from client
When adding dynamic data structure config from client, a check
for existing conflicting static config has to be made before
broadcasting the config to all members.

Also fixes an issue with CacheSimpleConfig copies not being
equal to their original which resulted in inability to add
the same cache config twice.