Fix cluster-name in examples and cover fully in tests
The commit addresses the following:
- full-example YAML config had an incorrect cluster name configuration
that was not caught by tests because the default cluster name was used
in the XML counterpart too. This is fixed and the cluster name changed
to non-default
- hazelcast-fullconfig-without-network xml and yaml files had the same
problem and the fix is the same.
- hazelcast-default.yaml had the same problem and it wasn't caught for
the same reason. The fix is the same unless the default cluster name is
kept.
- hazelcast-client-full and hazelcast-client-full-example had the same
problem and fixed the same way
- Specific tests were missing for cluster name in member config builder
tests.