Prevent migration operations running before previous finalization completes v2
If a node was the source of a backup replica migration
and then if it becomes the destination of another replica migration
on the same partition, `MigrationOperation` can be executed before
the former migration is finalized. Reason is, sources of backup migrations
are not part of migration transactions and they learn the migration
only while applying completed migrations.

Previously, when we detect this, we skipped the execution of previous
migrations finalization to prevent data loss.

But this may cause replica data leak, because of ignored finalization.

This time, migration finalizations are registered before
they are executed or enqueued. If there's an pending finalization
while a migration operation is being executed then migration operation
is rejected with a retryable exception.