Add metrics test rule
Introduce metrics test rule that records the metrics for every instance
periodically and retains the last N recording where N is configurable
and by default is 10. If there is a test failure, it prints the recorded
metrics for every instance to the stdout.

Every test extending `HazelcastTestSupport` and creating its test
Hazelcast instances with the factory methods provided by the support
class use the rule automatically. The other tests may use
`factory.withMetricsRule(metricsRule)` to leverage it.

An example output in case of a failure:

```

Metrics recorded during the test:

[charming_brahmagupta] [09:22:48,478]
[partitionId=19,unit=count,metric=operation.partition.executedOperationsCounter]=0
[charming_brahmagupta] [09:22:48,478]
[thread=hz.charming_brahmagupta.priority-generic-operation.thread-0,unit=count,metric=operation.thread.completedTotalCount]=0
[charming_brahmagupta] [09:22:48,478]
[thread=hz.charming_brahmagupta.generic-operation.thread-3,unit=count,metric=operation.thread.errorCount]=0
[charming_brahmagupta] [09:22:48,478]
[thread=hz.charming_brahmagupta.partition-operation.thread-0,unit=count,metric=operation.thread.completedOperationBatchCount]=0
[charming_brahmagupta] [09:22:48,478]
[partitionId=180,unit=count,metric=operation.partition.executedOperationsCounter]=0
...
```