Replace readData/writeData with readObject/writeObject (#14940)
Replace readData/writeData with readObject/writeObject

Replaces usage of obj.writeData(out) and obj.readData(in) with
in.readObject/out.writeObject. This reduces the probability of errors in
case the Versioned interface is introduced mistakenly on the object
being serialized but not the object performing the writeData/readData
call.

This PR does not cover all places where this is present, though. Using
writeObject/readObject will write two more ints per object in the
serialized format. In some cases (e.g. serialising many entries), this
may increase the size of the serialized format too much.

Fixes: https://github.com/hazelcast/hazelcast/issues/14909