Add metrics test rule (#16199)
Introduce metrics test rule that records the metrics for every instance
periodically and retains the last N recording where N is configurable
and by default is 10. If there is a test failure, it dumps the recorded
metrics for every Hazelcast instance to the stdout. The output can be
a bit verbose, if the test runs for seconds and uses multiple Hazelcast
instances, runs with default partition and thread pool count. The 
change makes the instances to collect, record and dump debug
metrics too. 

Every test extending `HazelcastTestSupport` and creating its test
Hazelcast instances with the factory methods provided by the support
class use the rule automatically. The other tests may use
`factory.withMetricsRule(metricsRule)` to leverage it.

An example output after a failure:

```

Metrics recorded during the test:

[charming_brahmagupta] [09:22:48,478]
[partitionId=19,unit=count,metric=operation.partition.executedOperationsCounter]=0
[charming_brahmagupta] [09:22:48,478]
[thread=hz.charming_brahmagupta.priority-generic-operation.thread-0,unit=count,metric=operation.thread.completedTotalCount]=0
[charming_brahmagupta] [09:22:48,478]
[thread=hz.charming_brahmagupta.generic-operation.thread-3,unit=count,metric=operation.thread.errorCount]=0
[charming_brahmagupta] [09:22:48,478]
[thread=hz.charming_brahmagupta.partition-operation.thread-0,unit=count,metric=operation.thread.completedOperationBatchCount]=0
[charming_brahmagupta] [09:22:48,478]
[partitionId=180,unit=count,metric=operation.partition.executedOperationsCounter]=0
...
```