Additional logging in case of failure in NioChannelMemoryLeakTest.testNioChannelLeak_afterMultipleSplitBrainMerges
Rationale: this commit extends the assertion in
assertNoChannelLeak() to see if both the NioChannel and
its socket channel are open/closed. If, for instance,
the socket channel is closed but the NioChannel is not,
the socket channel must have been closed by some other
means (an interrupt?) than AbstractChannel.close().

Relates to https://github.com/hazelcast/hazelcast/issues/14985