Expose entry data on WAN events
Exposes entry data on WAN publisher SPI events. Without this change,
there was no publicly available method which would return the entry data
which effectively makes the SPI unusable as the only publicly available
data was service and object name. We now additionally expose event type
(added/updated, synced or removed) and WAN event object. The object is
an usually returned as an Object but is either the key on which
deletion occurred or one of two public interfaces - EntryView (for map)
or CacheEntryView (for cache) with lazy deserialisation.

Also, moved the getServiceName method to private API. It would return a
String service name which would then force the user to check it against
private API - either MapService or CacheService. Instead, we move
provide another method which returns an public enum.