REST API cleanup, part 2
- all HTTP status codes, including HTTP 200 OK can now return a response
body
- moved helper methods from HttpPostCommandProcessor to HttpCommand
- removed the exception handling and authentication helper methods. The
exception handling is now done on one place and not by each handler
independently. The authentication helper method wasn't that great since
it required constructing lambdas and required indentation of the handler
code which was not worth the effort
- the exception handling now always returns a HTTP 500 for an error with
a JSON body containing the exception message
- the authentication and parameter parsing is now handled by the method
decodeParamsAndAuthenticate. This method will also expect a number of
parameters. If the request does not contain at least that many
parameters, HTTP 400 is returned with a JSON describing the requested
number of parameters
- HTTP 400 is NOT returned now if any handler throws an
IndexOutOfBoundsException. The assumption was that this was probably
caused by a bad request but that doesn't need to be the case.
- added credentials checking to POST methods for map and queue. This is
debatable but I assumed that mutation methods should be authenticated
(the DELETE method isn't but it requires a bigger change to add request
data to HttpDeleteCommand)
- added some javadoc, Nonnull and Nullable annotations and some minor
cleanups

Possible future changes:
- add parameter validations and return HTTP 400 on bad parameters
- return HTTP 503 or HTTP 501 when EE handlers are used on OS side
instead of returning HTTP 200 or HTTP 500 (examples - setting license
key, WAN REST endpoints)