Make CP group ids unique between multiple inits of CP Subsystem
We have 2 fields in CPGroupId: seed and commitIndex. When a new CP group is
created, commit index of its CP group id denotes on which commit of
the Metadata CP group this custom CP group is created. In addition to this,
the seed field is incremented when CP Subsystem is reset. By this way, we
ensure uniqueness of <seed, commit index> pairs of CP group ids. However, if
the whole Hazelcast cluster is restarted, the seed will be 0 again. This can
cause the Default CP group id to be same with the one in the previous Hazelcast
cluster because usually the Default CP group is created on the same Metadata CP
group commit index (=5). If this happens, CP proxies that are created for
the Default CP group of the previous Hazelcast cluster can continue to talk to
the Default CP group created on the new Hazelcast cluster, which means we lose
our consistency guarantees.

To prevent this problem, we don't rely on the Metadata CP group commit index
for generation of CP group ids. Instead, we generate random numbers for
the commit index field of CPGroupId.

Fixes #16221