REST API cleanup, part 2 (#16148)
REST API cleanup, part 2

- all HTTP status codes, including HTTP 200 OK can now return a response
body
- moved helper methods from HttpPostCommandProcessor to HttpCommand
- removed the exception handling and authentication helper methods. The
exception handling is now done on one place and not by each handler
independently. The authentication helper method wasn't that great since
it required constructing lambdas and required indentation of the handler
code which was not worth the effort
- the exception handling now always returns a HTTP 500 for an error with
a JSON body containing the exception message
- the authentication and parameter parsing is now handled by the method
decodeParamsAndAuthenticate. This method will also expect a number of
parameters. If the request does not contain at least that many
parameters, HTTP 400 is returned with a JSON describing the requested
number of parameters
- HTTP 400 is NOT returned now if any handler throws an
IndexOutOfBoundsException. The assumption was that this was probably
caused by a bad request but that doesn't need to be the case.
- added some javadoc, Nonnull and Nullable annotations and some minor
cleanups
- added HttpBadRequestException and HttpForbiddenException and throwing 
these from the parameter parsing method, catching them in 
HttpPostCommandProcessor#handle and sending responses
- added test to test both of these for all POST URI's that require authentication. 
As a consequence, removed some existing (now duplicate) tests that were 
testing for FORBIDDEN. I didn't add tests to assert behaviour of correct 
(authenticated) requests as I assume there are existing tests for that.
- added more assertions to assert JSON output on some POST tests that only
 asserted http status code
- removed authentication from map/queue POST URIs. It wasn't there before 
this PR and introducing it also brought some issues since the handler methods
expect that there are no String parameters but instead use the entire request
data as a value for the map/queue. Introducing authentication as parameters 
complicates this.