Revert "issues/13540: Transaction Propagation support implementation (#15141)"
This reverts commit 51675d817b8fbe627045d4acbab0e55f6fc56656.

After BETA-1 release, we have encoutered issues with nested transactions
(e.g. suspend / resume logic with Weblogic, JCA adaptor) and it seems
the feature is incomplete. We are therefore reverting to prevent giving
a false feeling of confidence. We need to invest more time into the
development of the feature or, actually, we need to redesign
transactions from the ground up.