Expect anti-entropy requests in WanPublisherMigrationTest assertions
In addition to migration process events,
WanMigrationAwarePublisher.prepareEventContainerReplicationData and
WanMigrationAwarePublisher.processEventContainerReplicationData can
also be called when the replication mechanism detects that there
is a difference between the partition owner and backup replica
and there's an anti-entropy process. Even if we disable periodic checks
for partition replica anti-entropy, this process can also be triggered
by reordered backup events. Because of this, we weaken the assertion for
processed WAN replication operations to allow for anti-entropy events.

Fixes: https://github.com/hazelcast/hazelcast/issues/16308