Use WanMapEntryView and WanCacheEntryView in WAN code and expose to user
Uses WanMapEntryView and WanCacheEntryView both as internal copies of
entry data for WAN replication and allows exposing the views to users
through public API. The classes allow for lazy deserialization but
internally we use serialized instances of entry data.

These classes were already used in WAN code but now we expand their use
even in WAN replication entry points. This way, we avoid an intermediary
SimpleEntryView which was created in map operations only to be converted
into a WAN*EntryView shortly after.

Moved WanCacheEntryView from EE into OS because of this.