Use explicit port on CloudDiscovery for private address
In certain scenarios like AWS private link, public and private
addresses can not have same port. Let the client use the explicit
private address if it is provided.
If it is not given explicitly, the client will use the port of public
address for the private address as before to maintain backward
compatibility.