Expose entry data on WAN events (#16228)
Expose entry data on WAN events

Exposes entry data on WAN publisher SPI events. Without this change,
there was no publicly available method which would return the entry data
which effectively makes the SPI unusable as the only publicly available
data was service and object name. We now additionally expose event type
(added/updated, synced or removed) and WAN event object. The object is
an usually returned as an Object but is either the key on which
deletion occurred or one of two public interfaces - WanMapEntryView (for map)
or WanCacheEntryView (for cache) with lazy deserialisation.

Uses WanMapEntryView and WanCacheEntryView both as internal copies of
entry data for WAN replication and allows exposing the views to users
through public API. The classes allow for lazy deserialization but
internally we use serialized instances of entry data.

These classes were already used in WAN code but now we expand their use
even in WAN replication entry points. This way, we avoid an intermediary
SimpleEntryView which was created in map operations only to be converted
into a WAN*EntryView shortly after.

Moved WanCacheEntryView from EE into OS because of this.