Factory to populate Config object from well-known locations
It allows to load config from an external file and then customize
it programmatically. This is useful when embedding Hazelcast into
managed environments: When a runtime container wants to create an
instance as if it was created via `Hazelcast.newHazelcastInstance()`,
but it wants to inject an instance of `ManagedContext` into Hazelcast
configuration.

Without this capability all containers have to duplicate the `Config`
location strategy: "First try this, then that, etc."

Related to https://github.com/spring-projects/spring-boot/issues/19487

TODO: Tests