Remove redundant IsFailoverSupportedMessage
After client has authenticated, we were sending
IsFailoverSupportedMessageTask (when newHazelcastFailoverClient is used)
to query if members have the support.

In this pr, we put this info to Authentication response
to avoid second step.

This was a different message task for backward compatibility reasons.
When client has connected to a member that feature is not available yet,
we wanted to get exception. Since there is no old server anymore,
a second message task is not needed.

related protocol change https://github.com/hazelcast/hazelcast-client-protocol/pull/288
fixes https://github.com/hazelcast/hazelcast/issues/16336