Handle closed client connection after registration
It could happen that this connection is already closed and onConnectionClose()
is called even before the authentication handling logic is executed. In this
case, now we have a closed but registered connection in the client connection
manager. We do a final check to remove this connection if needed.

This check was already here before and removed with 3d897c278 because of wrong
reasoning.

Fixes #16288