Removed the dependencies on Data from the SplitBrainMergePolicy API
Changed MergingValue/MergingEntry so that the newly introduced
methods getRawValue/Key (which supersede the old getValue/Key)
return the in-memory representation as Object. The deserialized
value can be obtained using getDeserializedValue/Key.

The merge types in SplitBrainMergeType no longer depend on Data.
Instead of specifying the in-memory storage type of the keys and
values (Data), they specify the deserialized types. Also, the value
type has been removed from the various 'view' interfaces such as
MergingHits, MergingCreationTime, etc.

The generic type signature of SplitBrainMergePolicy has changed
to specify the (deserialized) type of the merging value; the
type of merging value object (typically, MergingValue<V>, but it can
be a composition like MergingEntry<String, V> & MergingHits); and
finally, the return type merge() method (typically Object for merge
policies that work with the in-memory storage type).