Make Ringbuffer not throw StaleSequenceException on read-many operation (#16303)
The problem we want to fix is that when producers are fast, consumers can have a really hard time keeping- and catching up with them (they often fail to do so, even when they would be able to process fast enough). For example if your head is at 1, and you ask for 100 items from 0, the Ringbuffer.readManyAsync call just throws an exception, instead of returning what’s available. By the time the exception travels to the consumer, gets processed and a new request is made, the requested sequence number tends to be stale again and will result in yet another exception. Meanwhile data is being overwritten in the Ringbuffer and is lost to the consumer.

The solution aims to remove the StaleSequenceException being thrown by the ReadManyOperation in such situations and just have it return the data that's available.

We are able to do this for Ringbuffer.readManyAsync, because it returns a ReadResultSet carrying sequence numbers, thus allowing the client to notice potential sequence gaps and decide if it can tolerate them or not. On the other had we can't do the same for Ringbuffer.readOne because there is no way to observe sequence numbers in that case (it returns just an element from the Ringbuffer).