Fix LossToleranceTest
Because of the behaviour change introduced in
https://github.com/hazelcast/hazelcast/pull/16303, when the requested
sequence is larger than the largest sequence (tailSequence) + 1, we
don't listen from the oldest sequence (headSequence) but rather from the
tailSequence + 1. Both approaches are fine and both approaches work
better in some scenarios. Since the listener is loss tolerant, we can
skip items from headSequence..tailSequence+1 anyway.

Fixed the test to adhere to the new behaviour. We assume that eventually
as we publish an item, it will reach the listener.

A better fix would be to introduce unique IDs per ringbuffer, where we
would then be able to distinguish between a completely lost ringbuffer
and a ringbuffer which has not received the last few items and
appropriately reset the requested sequence to the headSequence or
tailSequence.

Fixes: https://github.com/hazelcast/hazelcast/issues/16430