Fix ClientRegressionWithRealNetworkTest failure
If the first CREATED event is missed after the client reconnects to the new
cluster, the assertion at the end of the test fails, hence prevents the remove
call to be executed. So map.remove() is never executed and all other puts fire
UPDATED events. map.remove() call is moved before map.put() to fire CREATED
events in all cases.

Special thanks to Sancar and Ihsan for their great help on digging the issue.

Fixes #16126