Fail MigrationOperation when TargetNotMemberException is thrown
`MigrationRequestOperation` checks for existence of migration destination
in `verifyExistingDestination()` method and fails the migration
with `TargetNotMemberException` if destination doesn't exist.

But if destination leaves the cluster while `MigrationOperation` is being invoked,
`MigrationOperation` is retried with `TargetNotMemberException` until retries timeout.

`MigrationOperation` should fail when `TargetNotMemberException` is thrown
after it's invoked, because it's known that if `MigrationOperation` is invoked
destination was there but left in the meantime.

(cherry picked from commit 367e2cafbc407bfa2520af424ed8e6bc33802b8e)