Fix client listener registry leaks
Background:
When a listener is added from an endpoint, we also register a
destroyAction. A destroyAction is a function to deregister the listener.
It is used when endpoint is removed. When endpoint is removed, we call
 all the registered destroy actions.

Racy scenario is as follows:
1. A listener is added, but not registered the destroy action.
2. Endpoint is removed because the client is disconnected.
All the destroy actions currently registered are called.
3. DestroyAction is registered to endpoint after it is destroyed.
4. After this point, there is no one to call the last destroyAction,
hence the leak.

As fix, we have added a second check if endpoint is destroyed after
a destroy action is put.

fixes https://github.com/hazelcast/hazelcast/issues/16429